chore: remove passthrough component
